+++ source/GameInterface/Services/MobileParties/Handlers/MobilePartyDataHandler.cs
@@ -31,9 +31,6 @@ internal class MobilePartyDataHandler : IHandler
         messageBroker.Subscribe<ChangePartyArmy>(Handle_ChangePartyArmy);
         messageBroker.Subscribe<PartyComponentChanged>(Handle_PartyComponentChanged);
         messageBroker.Subscribe<NetworkChangePartyComponent>(Handle_ChangePartyComponent);
-
-        messageBroker.Subscribe<ActualClanChanged>(Handle_ActualClanChanged);
-        messageBroker.Subscribe<NetworkChangeActualClan>(Handle_ChangeActualClan);
     }
 
     public void Dispose()
@@ -41,9 +38,6 @@ internal class MobilePartyDataHandler : IHandler
         messageBroker.Unsubscribe<ChangePartyArmy>(Handle_ChangePartyArmy);
         messageBroker.Unsubscribe<PartyComponentChanged>(Handle_PartyComponentChanged);
         messageBroker.Unsubscribe<NetworkChangePartyComponent>(Handle_ChangePartyComponent);
-
-        messageBroker.Unsubscribe<ActualClanChanged>(Handle_ActualClanChanged);
-        messageBroker.Unsubscribe<NetworkChangeActualClan>(Handle_ChangeActualClan);
     }
 
     private void Handle_ChangePartyArmy(MessagePayload<ChangePartyArmy> payload)
@@ -96,42 +90,5 @@ internal class MobilePartyDataHandler : IHandler
         objectManager.Remove(componentId);
 
         party._partyComponent = component;
-    }
-
-    private void Handle_ActualClanChanged(MessagePayload<ActualClanChanged> payload)
-    {
-        var message = new NetworkChangeActualClan(
-            payload.What.PartyId,
-            payload.What.ClanId
-        );
-
-        network.SendAll(message);
-    }
-
-    private void Handle_ChangeActualClan(MessagePayload<NetworkChangeActualClan> payload)
-    {
-        var partyId = payload.What.PartyId;
-        var clanId = payload.What.ClanId;
-
-        if (objectManager.TryGetObject(partyId, out MobileParty party) == false)
-        {
-            Logger.Error("Failed to find party with stringId {stringId}", partyId);
-            return;
-        }
-
-        // Set clan to null if id was null
-        if (clanId == null)
-        {
-            ActualClanPatches.OverrideSetActualClan(party, null);
-            return;
-        }
-
-        if (objectManager.TryGetObject(clanId, out Clan clan) == false)
-        {
-            Logger.Error("Failed to find Clan with stringId {stringId}", clanId);
-            return;
-        }
-
-        ActualClanPatches.OverrideSetActualClan(party, clan);
     }
 }
