--- source/GameInterface/Services/MobileParties/Messages/ActualClanChanged.cs
@@ -1,15 +1,0 @@
-﻿using Common.Messaging;
-using ProtoBuf;
-
-namespace GameInterface.Services.MobileParties.Messages;
-internal class ActualClanChanged : IEvent
-{
-    public string PartyId { get; }
-    public string ClanId { get; }
-
-    public ActualClanChanged(string partyId, string clanId)
-    {
-        PartyId = partyId;
-        ClanId = clanId;
-    }
-}
