+++ source/GameInterface/Services/MobileParties/Patches/MobilePartyPropertyPatches.cs
@@ -1,7 +1,6 @@
 ﻿using Common.Logging;
 using Common.Messaging;
 using GameInterface.Policies;
-using GameInterface.Services.MobileParties.Messages;
 using HarmonyLib;
 using Serilog;
 using System;
@@ -27,6 +26,13 @@ public class MobilePartyPropertyPatches
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
 
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
+
         var message = new MobilePartyPropertyChanged(PropertyType.Army, __instance.StringId, value?.LeaderParty.StringId);
         MessageBroker.Instance.Publish(__instance, message);
 
@@ -38,6 +44,13 @@ public class MobilePartyPropertyPatches
     private static bool SetCustomNamePrefix(MobileParty __instance, TextObject value)
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
+
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
 
         var message = new MobilePartyPropertyChanged(PropertyType.CustomName, __instance.StringId, value?.Value);
         MessageBroker.Instance.Publish(__instance, message);
@@ -51,6 +64,13 @@ public class MobilePartyPropertyPatches
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
 
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
+
         var message = new MobilePartyPropertyChanged(PropertyType.LastVisitedSettlement, __instance.StringId, value?.StringId);
         MessageBroker.Instance.Publish(__instance, message);
 
@@ -62,6 +82,13 @@ public class MobilePartyPropertyPatches
     private static bool SetAggressivenessPrefix(MobileParty __instance, float value)
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
+
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
 
         var message = new MobilePartyPropertyChanged(PropertyType.Aggressiveness, __instance.StringId, value.ToString());
         MessageBroker.Instance.Publish(__instance, message);
@@ -75,6 +102,13 @@ public class MobilePartyPropertyPatches
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
 
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
+
         var message = new MobilePartyPropertyChanged(PropertyType.ArmyPositionAdder, __instance.StringId, value.x.ToString(), value.y.ToString());
         MessageBroker.Instance.Publish(__instance, message);
 
@@ -87,42 +121,69 @@ public class MobilePartyPropertyPatches
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
 
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
+
         var message = new MobilePartyPropertyChanged(PropertyType.Objective, __instance.StringId, value.ToString());
         MessageBroker.Instance.Publish(__instance, message);
 
         return ModInformation.IsServer;
     }
 
-    //[HarmonyPatch(nameof(MobileParty.Ai), MethodType.Setter)]
-    //[HarmonyPrefix]
-    //private static bool SetAiPrefix(MobileParty __instance, MobilePartyAi value)
-    //{
-    //    if (CallOriginalPolicy.IsOriginalAllowed()) return true;
+    [HarmonyPatch(nameof(MobileParty.Ai), MethodType.Setter)]
+    [HarmonyPrefix]
+    private static bool SetAiPrefix(MobileParty __instance, MobilePartyAi value)
+    {
+        if (CallOriginalPolicy.IsOriginalAllowed()) return true;
 
-    //    //Is this one needed?
-    //    var message = new MobilePartyPropertyChanged(PropertyType.Ai, __instance.StringId, value.ToString());
-    //    MessageBroker.Instance.Publish(__instance, message);
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
 
-    //    return ModInformation.IsServer;
-    //}
+        var message = new MobilePartyPropertyChanged(PropertyType.Ai, __instance.StringId, value.ToString());
+        MessageBroker.Instance.Publish(__instance, message);
 
-    //[HarmonyPatch(nameof(MobileParty.Party), MethodType.Setter)]
-    //[HarmonyPrefix]
-    //private static bool SetPartyPrefix(MobileParty __instance, PartyBase value)
-    //{
-    //    if (CallOriginalPolicy.IsOriginalAllowed()) return true;
+        return ModInformation.IsServer;
+    }
 
-    //    var message = new MobilePartyPropertyChanged(PropertyType.Party, __instance.StringId, value.MobileParty?.StringId);
-    //    MessageBroker.Instance.Publish(__instance, message);
+    [HarmonyPatch(nameof(MobileParty.Party), MethodType.Setter)]
+    [HarmonyPrefix]
+    private static bool SetPartyPrefix(MobileParty __instance, PartyBase value)
+    {
+        if (CallOriginalPolicy.IsOriginalAllowed()) return true;
 
-    //    return ModInformation.IsServer;
-    //}
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(PartyBase), Environment.StackTrace);
+            return true;
+        }
+
+        var message = new MobilePartyPropertyChanged(PropertyType.Party, __instance.StringId, value.MobileParty?.StringId);
+        MessageBroker.Instance.Publish(__instance, message);
+
+        return ModInformation.IsServer;
+    }
 
     [HarmonyPatch(nameof(MobileParty.IsActive), MethodType.Setter)]
     [HarmonyPrefix]
     private static bool SetIsActivePrefix(MobileParty __instance, bool value)
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
+
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
 
         var message = new MobilePartyPropertyChanged(PropertyType.IsActive, __instance.StringId, value.ToString());
         MessageBroker.Instance.Publish(__instance, message);
@@ -136,30 +197,31 @@ public class MobilePartyPropertyPatches
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
 
-        //Is this needed?
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
+
         var message = new MobilePartyPropertyChanged(PropertyType.ThinkParamsCache, __instance.StringId, value.ToString());
         MessageBroker.Instance.Publish(__instance, message);
 
         return ModInformation.IsServer;
     }
 
-    //[HarmonyPatch(nameof(MobileParty.ShortTermBehavior), MethodType.Setter)]
-    //[HarmonyPrefix]
-    //private static bool SetShortTermBehaviorPrefix(MobileParty __instance, AiBehavior value)
-    //{
-    //    if (CallOriginalPolicy.IsOriginalAllowed()) return true;
-
-    //    var message = new MobilePartyPropertyChanged(PropertyType.ShortTermBehaviour, __instance.StringId, value.ToString());
-    //    MessageBroker.Instance.Publish(__instance, message);
-
-    //    return ModInformation.IsServer;
-    //}
-
     [HarmonyPatch(nameof(MobileParty.IsPartyTradeActive), MethodType.Setter)]
     [HarmonyPrefix]
     private static bool SetIsPartyTradeActivePrefix(MobileParty __instance, bool value)
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
+
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
 
         var message = new MobilePartyPropertyChanged(PropertyType.IsPartyTradeActive, __instance.StringId, value.ToString());
         MessageBroker.Instance.Publish(__instance, message);
@@ -173,6 +235,13 @@ public class MobilePartyPropertyPatches
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
 
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
+
         var message = new MobilePartyPropertyChanged(PropertyType.PartyTradeGold, __instance.StringId, value.ToString());
         MessageBroker.Instance.Publish(__instance, message);
 
@@ -184,6 +253,13 @@ public class MobilePartyPropertyPatches
     private static bool SetPartyTradeTaxGoldPrefix(MobileParty __instance, int value)
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
+
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
 
         var message = new MobilePartyPropertyChanged(PropertyType.PartyTradeTaxGold, __instance.StringId, value.ToString());
         MessageBroker.Instance.Publish(__instance, message);
@@ -199,6 +275,13 @@ public class MobilePartyPropertyPatches
     //{
     //    if (CallOriginalPolicy.IsOriginalAllowed()) return true;
 
+    //    if (ModInformation.IsClient)
+    //    {
+    //        Logger.Error("Client created unmanaged {name}\n"
+    //            + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+    //        return true;
+    //    }
+
     //    var message = new MobilePartyPropertyChanged(PropertyType.StationaryStartTime, __instance.StringId, value.NumTicks.ToString());
     //    MessageBroker.Instance.Publish(__instance, message);
 
@@ -210,6 +293,13 @@ public class MobilePartyPropertyPatches
     private static bool SetVersionNoPrefix(MobileParty __instance, int value)
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
+
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
 
         var message = new MobilePartyPropertyChanged(PropertyType.VersionNo, __instance.StringId, value.ToString());
         MessageBroker.Instance.Publish(__instance, message);
@@ -223,6 +313,13 @@ public class MobilePartyPropertyPatches
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
 
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
+
         var message = new MobilePartyPropertyChanged(PropertyType.ShouldJoinPlayerBattles, __instance.StringId, value.ToString());
         MessageBroker.Instance.Publish(__instance, message);
 
@@ -234,6 +331,13 @@ public class MobilePartyPropertyPatches
     private static bool SetIsDisbandingPrefix(MobileParty __instance, bool value)
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
+
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
 
         var message = new MobilePartyPropertyChanged(PropertyType.IsDisbanding, __instance.StringId, value.ToString());
         MessageBroker.Instance.Publish(__instance, message);
@@ -247,6 +351,13 @@ public class MobilePartyPropertyPatches
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
 
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
+
         var message = new MobilePartyPropertyChanged(PropertyType.CurrentSettlement, __instance.StringId, value?.StringId);
         MessageBroker.Instance.Publish(__instance, message);
 
@@ -258,6 +369,13 @@ public class MobilePartyPropertyPatches
     private static bool SetAttachedToPrefix(MobileParty __instance, MobileParty value)
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
+
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
 
         var message = new MobilePartyPropertyChanged(PropertyType.AttachedTo, __instance.StringId, value?.StringId);
         MessageBroker.Instance.Publish(__instance, message);
@@ -271,6 +389,13 @@ public class MobilePartyPropertyPatches
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
 
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
+
         var message = new MobilePartyPropertyChanged(PropertyType.BesiegerCamp, __instance.StringId, value?.ToString());
         MessageBroker.Instance.Publish(__instance, message);
 
@@ -282,6 +407,13 @@ public class MobilePartyPropertyPatches
     private static bool SetScoutPrefix(MobileParty __instance, Hero value)
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
+
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
 
         var message = new MobilePartyPropertyChanged(PropertyType.Scout, __instance.StringId, value?.StringId);
         MessageBroker.Instance.Publish(__instance, message);
@@ -295,6 +427,13 @@ public class MobilePartyPropertyPatches
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
 
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
+
         var message = new MobilePartyPropertyChanged(PropertyType.Engineer, __instance.StringId, value?.StringId);
         MessageBroker.Instance.Publish(__instance, message);
 
@@ -306,6 +445,13 @@ public class MobilePartyPropertyPatches
     private static bool SetQuartermasterPrefix(MobileParty __instance, Hero value)
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
+
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
 
         var message = new MobilePartyPropertyChanged(PropertyType.Quartermaster, __instance.StringId, value?.StringId);
         MessageBroker.Instance.Publish(__instance, message);
@@ -319,29 +465,57 @@ public class MobilePartyPropertyPatches
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
 
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
+
         var message = new MobilePartyPropertyChanged(PropertyType.Surgeon, __instance.StringId, value?.StringId);
         MessageBroker.Instance.Publish(__instance, message);
 
         return ModInformation.IsServer;
     }
 
-    //[HarmonyPatch(nameof(MobileParty.ActualClan), MethodType.Setter)]
-    //[HarmonyPrefix]
-    //private static bool SetActualClanPrefix(MobileParty __instance, Clan value)
-    //{
-    //    if (CallOriginalPolicy.IsOriginalAllowed()) return true;
+    [HarmonyPatch(nameof(MobileParty.ActualClan), MethodType.Setter)]
+    [HarmonyPrefix]
+    private static bool SetActualClanPrefix(MobileParty __instance, Clan value)
+    {
+        if (CallOriginalPolicy.IsOriginalAllowed()) return true;
 
-    //    var message = new MobilePartyPropertyChanged(PropertyType.ActualClan, __instance.StringId, value?.StringId);
-    //    MessageBroker.Instance.Publish(__instance, message);
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
 
-    //    return ModInformation.IsServer;
-    //}
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(Clan), Environment.StackTrace);
+            return true;
+        }
+
+        var message = new MobilePartyPropertyChanged(PropertyType.ActualClan, __instance.StringId, value?.StringId);
+        MessageBroker.Instance.Publish(__instance, message);
+
+        return ModInformation.IsServer;
+    }
 
     [HarmonyPatch(nameof(MobileParty.RecentEventsMorale), MethodType.Setter)]
     [HarmonyPrefix]
     private static bool SetRecentEventsMoralePrefix(MobileParty __instance, float value)
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
+
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
 
         var message = new MobilePartyPropertyChanged(PropertyType.RecentEventsMorale, __instance.StringId, value.ToString());
         MessageBroker.Instance.Publish(__instance, message);
@@ -355,29 +529,50 @@ public class MobilePartyPropertyPatches
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
 
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
+
         var message = new MobilePartyPropertyChanged(PropertyType.EventPositionAdder, __instance.StringId, value.X.ToString(), value.Y.ToString());
         MessageBroker.Instance.Publish(__instance, message);
 
         return ModInformation.IsServer;
     }
 
-    //[HarmonyPatch(nameof(MobileParty.IsInspected), MethodType.Setter)]
-    //[HarmonyPrefix]
-    //private static bool SetIsInspectedPrefix(MobileParty __instance, bool value)
-    //{
-    //    if (CallOriginalPolicy.IsOriginalAllowed()) return true;
+    [HarmonyPatch(nameof(MobileParty.IsInspected), MethodType.Setter)]
+    [HarmonyPrefix]
+    private static bool SetIsInspectedPrefix(MobileParty __instance, bool value)
+    {
+        if (CallOriginalPolicy.IsOriginalAllowed()) return true;
 
-    //    var message = new MobilePartyPropertyChanged(PropertyType.IsInspected, __instance.StringId, value.ToString());
-    //    MessageBroker.Instance.Publish(__instance, message);
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
 
-    //    return ModInformation.IsServer;
-    //}
+        var message = new MobilePartyPropertyChanged(PropertyType.IsInspected, __instance.StringId, value.ToString());
+        MessageBroker.Instance.Publish(__instance, message);
+
+        return ModInformation.IsServer;
+    }
 
     [HarmonyPatch(nameof(MobileParty.MapEventSide), MethodType.Setter)]
     [HarmonyPrefix]
     private static bool SetMapEventSidePrefix(MobileParty __instance, MapEventSide value)
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
+
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
 
         var message = new MobilePartyPropertyChanged(PropertyType.MapEventSide, __instance.StringId, value.ToString());
         MessageBroker.Instance.Publish(__instance, message);
@@ -391,6 +586,13 @@ public class MobilePartyPropertyPatches
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
 
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
+
         var message = new MobilePartyPropertyChanged(PropertyType.PartyComponent, __instance.StringId, value.ToString());
         MessageBroker.Instance.Publish(__instance, message);
 
@@ -402,6 +604,13 @@ public class MobilePartyPropertyPatches
     private static bool SetIsMilitaPrefix(MobileParty __instance, bool value)
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
+
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
 
         var message = new MobilePartyPropertyChanged(PropertyType.IsMilita, __instance.StringId, value.ToString());
         MessageBroker.Instance.Publish(__instance, message);
@@ -415,6 +624,13 @@ public class MobilePartyPropertyPatches
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
 
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
+
         var message = new MobilePartyPropertyChanged(PropertyType.IsLordParty, __instance.StringId, value.ToString());
         MessageBroker.Instance.Publish(__instance, message);
 
@@ -426,6 +642,13 @@ public class MobilePartyPropertyPatches
     private static bool SetIsVillagerPrefix(MobileParty __instance, bool value)
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
+
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
 
         var message = new MobilePartyPropertyChanged(PropertyType.IsVillager, __instance.StringId, value.ToString());
         MessageBroker.Instance.Publish(__instance, message);
@@ -439,6 +662,13 @@ public class MobilePartyPropertyPatches
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
 
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
+
         var message = new MobilePartyPropertyChanged(PropertyType.IsCaravan, __instance.StringId, value.ToString());
         MessageBroker.Instance.Publish(__instance, message);
 
@@ -450,6 +680,13 @@ public class MobilePartyPropertyPatches
     private static bool SetIsGarrisonPrefix(MobileParty __instance, bool value)
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
+
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
 
         var message = new MobilePartyPropertyChanged(PropertyType.IsGarrison, __instance.StringId, value.ToString());
         MessageBroker.Instance.Publish(__instance, message);
@@ -463,6 +700,13 @@ public class MobilePartyPropertyPatches
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
 
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
+
         var message = new MobilePartyPropertyChanged(PropertyType.IsCustomParty, __instance.StringId, value.ToString());
         MessageBroker.Instance.Publish(__instance, message);
 
@@ -474,6 +718,13 @@ public class MobilePartyPropertyPatches
     private static bool SetIsBanditPrefix(MobileParty __instance, bool value)
     {
         if (CallOriginalPolicy.IsOriginalAllowed()) return true;
+
+        if (ModInformation.IsClient)
+        {
+            Logger.Error("Client created unmanaged {name}\n"
+                + "Callstack: {callstack}", typeof(MobilePartyAi), Environment.StackTrace);
+            return true;
+        }
 
         var message = new MobilePartyPropertyChanged(PropertyType.IsBandit, __instance.StringId, value.ToString());
         MessageBroker.Instance.Publish(__instance, message);
